pyproject.toml
[project]
name = "sae-qwen"
version = "0.1.0"
description = "Sparse Autoencoder training on Qwen2.5-1.5B activations"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "transformer-lens>=2.0.0",
    "transformers>=4.40.0",
    "datasets>=2.18.0",
    "wandb>=0.16.0",
    "einops>=0.7.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
    "accelerate>=0.27.0",
]

[project.optional-dependencies]
analysis = [
    "jupyter>=1.0.0",
    "matplotlib>=3.8.0",
    "plotly>=5.18.0",
    "pandas>=2.2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]  # Script-based project, no package structure

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
